requirements.txt
rasterio
geopandas
numpy
opencv-python
pillow
easydict